energy

+++ vla-scripts/energy/energy_model.py
@@ -203,8 +203,8 @@ class EnergyModel(nn.Module):
         
 
         # return E
-        context_mapped = self.state_linear(hN).to(torch.bfloat16)  # [B,S,Dh]
-        action_mapped  = self.pe_layer(self.action_linear(a)).to(torch.bfloat16)  # [B,H,Da]
+        context_mapped = self.state_linear(hN)  # [B,S,Dh]
+        action_mapped  = self.pe_layer(self.action_linear(a))  # [B,H,Da]
 
         # energy_feat = self.energy_bc(query=action_mapped.transpose(0, 1),
         #     value=context_mapped.transpose(0, 1),
